refactor: History controller logic
