--- Backend/sentinel.ai.api/Hubs/VideoFeedHub.cs
@@ -1,12 +1,0 @@
-using Microsoft.AspNetCore.SignalR;
-
-namespace sentinel.ai.infrastructure.Hubs
-{
-    public class VideoFeedHub : Hub
-    {
-        public async Task SendVideoFeed(byte[] videoData)
-        {
-            await Clients.All.SendAsync("ReceiveVideoFeed", videoData);
-        }
-    }
-}
