--- Backend/sentinel.ai.domain/Interfaces/IHistoryRepo.cs
@@ -1,11 +1,0 @@
-using sentinel.ai.domain.Dto;
-
-namespace sentinel.ai.domain.interfaces
-{
-    public interface IHistoryRepo
-    {
-        bool Insert(BehaviouralAssessment assessment);
-        Task<IEnumerable<BehaviouralAssessment>> GetAll();
-        Task UpdateLatest(string verdict, double confidence);
-    }
-}
