+++ Backend/sentinel.ai.domain/Services/IHistoryService.cs
@@ -1,0 +1,9 @@
+using sentinel.ai.domain.Dto;
+
+namespace sentinel.ai.domain.Services
+{
+    public interface IHistoryService
+    {
+        Task<IEnumerable<BehaviouralAssessment>> GetAll();
+    }
+}
