+++ Backend/sentinel.ai.infrastructure/Repositories/HistoricalDb.cs
@@ -1,15 +1,12 @@
-using System.Configuration;
-using System.Data;
-using System.Data.Common;
 using Dapper;
 using Microsoft.Extensions.Configuration;
 using MySql.Data.MySqlClient;
 using sentinel.ai.domain.Dto;
-using sentinel.ai.domain.interfaces;
+using sentinel.ai.domain.Repositories;
 
-namespace sentinel.ai.infrastructure.services;
+namespace sentinel.ai.infrastructure.Repositories;
 
-public class HistoricalDb : IHistoryRepo
+public class HistoricalDb : IHistoryRepository
 {
     private class DBConfig
     {
@@ -29,8 +26,10 @@ public class HistoricalDb : IHistoryRepo
             .GetRequiredSection("DatabaseConnection").Get<DBConfig>()
             ?? throw new InvalidOperationException("Invalid database connection configuration.");
 
-        string password = Environment.GetEnvironmentVariable("DATABASE__PASSWORD")
-            ?? throw new ConfigurationErrorsException("Database password not set in environment.");
+        // string password = Environment.GetEnvironmentVariable("DATABASE__PASSWORD")
+        //     ?? throw new ConfigurationErrorsException("Database password not set in environment.");
+
+        string password = "";
 
         _dbConnection = new MySqlConnection(
             $"Server={_dbConfig.Host};" +
@@ -46,7 +45,7 @@ public class HistoricalDb : IHistoryRepo
         return _dbConnection.QueryAsync<BehaviouralAssessment>($"SELECT * FROM {_dbConfig.HistoryTable};");
     }
 
-    public bool Insert(BehaviouralAssessment assessment)
+    public bool InsertHistory(BehaviouralAssessment assessment)
     {
         var command = new MySqlCommand($"INSERT INTO {_dbConfig.HistoryTable} (Timestamp, Image) VALUES (@timestamp, @imageBlob);", _dbConnection);
         command.Parameters.Add("@timestamp", MySqlDbType.DateTime).Value = assessment.Timestamp;
@@ -57,7 +56,7 @@ public class HistoricalDb : IHistoryRepo
         return result == 1;
     }
 
-    public Task UpdateLatest(string verdict, double confidence)
+    public void UpdateLatestHistory(string verdict, double confidence)
     {
         throw new NotImplementedException();
     }
