+++ Backend/sentinel.ai.infrastructure/Services/HistoryService.cs
@@ -1,0 +1,21 @@
+using sentinel.ai.domain.Dto;
+using sentinel.ai.domain.Services;
+
+namespace sentinel.ai.infrastructure.Services
+{
+    public class HistoryService : IHistoryService
+    {
+        public HistoryService()
+        {
+        }
+
+        public Task<IEnumerable<BehaviouralAssessment>> GetAll()
+        {
+            throw new NotImplementedException();
+        }
+    }
+
+    internal interface ILogger
+    {
+    }
+}
